part1_mock_interview/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.3.0
livekit-plugins-assemblyai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

